refactor: clean up External Access layer contracts (#69)
Move _transform_to_ansible_doc_format from galaxy.py to parser.py as a
public function (V-E3), and decouple parser.py from collections.py by
injecting collections_path as a parameter instead of importing it at
module level (V-L4). Update service-contracts.md to mark V-E2, V-E3,
and V-L4 as resolved.
